fix(hadron-acpi): correct hpet_number field type from u16 to u8
The ACPI spec defines "HPET Number" as a single byte (offset 52), but
HpetRaw declared it as u16. This made the struct 21 bytes instead of 20,
causing FromBytes::read_from() to fail with TruncatedData on QEMU's
standard 56-byte HPET table (36-byte header + 20 bytes data). The HPET
was never successfully detected.

### Fixed
- HpetRaw::hpet_number u16 → u8 to match ACPI spec (struct now 20 bytes)
- HpetTable::hpet_number u16 → u8 for consistency

### Changed
- HPET parse error now logs the actual error variant instead of generic
  "HPET not found"

+++ crates/acpi/src/hpet.rs
@@ -34,7 +34,7 @@ pub struct GenericAddress {
 struct HpetRaw {
     event_timer_block_id: u32,
     base_address: GenericAddress,
-    hpet_number: u16,
+    hpet_number: u8,
     minimum_tick: u16,
     page_protection: u8,
 }
@@ -47,7 +47,7 @@ pub struct HpetTable {
     /// Base address of the HPET register block.
     pub base_address: GenericAddress,
     /// HPET sequence number (used when multiple HPETs are present).
-    pub hpet_number: u16,
+    pub hpet_number: u8,
     /// Minimum clock tick in periodic mode, in femtoseconds.
     pub minimum_tick: u16,
     /// Page protection and OEM attribute.
